migrate types for route
