+++ types/src/index.ts
@@ -11,7 +11,20 @@ export * from "./front/data_source";
 export * from "./front/dataset";
 export * from "./front/document";
 export * from "./front/extract";
-export * from "./front/key";
+export * from "./front/membership_invitation";
+export * from "./front/plan";
+export * from "./front/project";
+export * from "./front/provider";
+export * from "./front/run";
+export * from "./front/assistant/agent";
+export * from "./front/assistant/actions/retrieval";
+export * from "./front/assistant/conversation";
+export * from "./front/assistant/actions/dust_app_run";
+export * from "./front/api_handlers/public/assistant";
+export * from "./front/api_handlers/public/data_sources";
+export * from "./front/api_handlers/internal/assistant";
+export * from "./front/api_handlers/internal/agent_configuration";
+
 export * from "./front/lib/actions/registry";
 export * from "./front/lib/actions/types";
 export * from "./front/lib/api/assistant/actions/dust_app_run";
